requirements.txt
foxglove-sdk
opencv-python
numpy
spatialmath-python
pin